[Chronik] Add chronik_info method into lib.rs
Summary: Lets us retrieve the `ChronikInfo` struct, which we can then extract the version from. The `assert_eq!()` needs a regex, as when the remote is updated, it will return a different String (with the same formatting which we can use for pattern matching).

Test Plan: `ninja-check-functional` && `cargo test` in `bitcoin-abc/modules/bitcoinsuite-chronik-client`

Reviewers: O1 Bitcoin ABC, #bitcoin_abc, Fabien

Reviewed By: O1 Bitcoin ABC, #bitcoin_abc, Fabien

Subscribers: Fabien

Differential Revision: https://reviews.bitcoinabc.org/D17528