[Cashtab] Disable claim button while we wait for server response
Summary:
Depends on D17559

After the user clicks to claim a reward, disable the button until we hear back from the server. Can take a few seconds to get the claim and we do not want multiple claims being made before we get a response.

Test Plan:
`npm test`, check test site at https://cashtab-local-dev.netlify.app/

I tried to catch this behavior in the tests but was not able to do it. The logic is pretty simple: disable the button in `handleClaim`, enable it if we get a response or an error.

Reviewers: #bitcoin_abc, emack

Reviewed By: #bitcoin_abc, emack

Differential Revision: https://reviews.bitcoinabc.org/D17560