[avalanche] Clear updates on each call to registerVotes
Summary: Since the API does not explicitly call for appending to updates, it is unintuitive for updates to not clear on each registerVotes call. This can lead to difficult to debug unit test failures where registerVotes appears to emit updates that were actually from past calls but never cleared.

Test Plan:
  ninja check-avalanche-processor_tests

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Subscribers: Fabien

Differential Revision: https://reviews.bitcoinabc.org/D17589