[avalanche] Check that avalanche invalidated items no longer poll in the dont_poll_invalid_item test case
Summary:
Technically the case of avalanche-triggered invalidation is already tested but it makes sense to re-test that behavior here to ensure avalanche invalidation is treated similarly as items being made invalid for other reasons.

This patch also improves the test coverage to actually check that the item is no longer being polled, not just that unexpected responses are handled correctly.

Depends on D17589

Test Plan:
  ninja check-avalanche-processor_tests

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Subscribers: Fabien

Differential Revision: https://reviews.bitcoinabc.org/D17590