[Cashtab] Do not wordbreak break-all for XECX summaries
Summary:
We use "break-all" for app actions as, most of the time, the content is unknown; it depends on what a user puts in a tx field. So we can have long strings like addresses that we need to "break-all" on, otherwise we get huge overflows that make the app look broken.

I have been seeing screenshots of "break-all" applied to XECX reward summaries which looks dumb. This is a hard-coded string and we do not need to worry about overflow issues.

So, do not "break all" for this specific app action. Add a prop that can control this behavior in other app actions as we uncover others that have this same issue.

Test Plan:
`npm test`, screenshots

before:
{F12936930}

after:
{F12936929}

Reviewers: #bitcoin_abc, emack

Reviewed By: #bitcoin_abc, emack

Differential Revision: https://reviews.bitcoinabc.org/D17628