[ecash-lib] [ecash-agora] Remove unneeded `ecc` parameters
Summary:
Now that we have a global `Ecc` instance, and `new Ecc` is basically a no-op, we can remove a lot of the `ecc: Ecc` params that are needed to use ecash-lib, simplifiying usage quite a bit.

The `TxBuilder.sign` function now receives a `params` object instead of a parameter list, which makes the code more readable.

This is a breaking change for both ecash-lib and ecash-agora, which we bundle in with the `initWasm` removal.

Depends on D17639.

Test Plan: CI

Reviewers: bytesofman, #bitcoin_abc

Reviewed By: bytesofman, #bitcoin_abc

Differential Revision: https://reviews.bitcoinabc.org/D17640