Re-enable the getavaaddr_interval_test
Summary: It appears that the check for relayable peers was done using the wrong availability score. Using the same name in different places for different things is incredibly confusing as demonstrated here. As opposed to the peer availability score which was unused, the node availability score is used under some circumstances like address relaying to filter out the obviously misbehaving nodes. The test was then not checking for the right condition due to using the wrong confusing score. Hopefully this fixes the test flakiness.

Test Plan:
  for i in {1..100}; do ./test/functional/test_runner.py abc_p2p_getavaaddr || break; done

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Subscribers: PiRK

Differential Revision: https://reviews.bitcoinabc.org/D17671