[avalanche] Deprecate the node availability score in the peerinfo RPC
Summary: This score is used internally by the node as a metric to filter out bad nodes in some cases. Unfortunately having it in the widely used peerinfo RPC caused the users to think it was a health metric for their node which it is not. Let's remove this field to avoid confusion, it was only added to help with testing.

Test Plan:
  ninja all check-all

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Subscribers: PiRK

Differential Revision: https://reviews.bitcoinabc.org/D17674