Improve reliability of abc_rpc_getavalancheproofs
Summary: This test can benefit from a noban rule to get a nice speedup and make the duration more repeatable. Also increase the finalization timeout for good measure as it managed to fail while not misbehaving on the macos CI agent that has the slowest networking over our machines.

Test Plan:
  ./test/functional/test_runner.py abc_rpc_getavalancheproofs

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Differential Revision: https://reviews.bitcoinabc.org/D17677