[Cashtab] Make sure whitelisted agora tokens load cache first
Summary:
If a user loads the agora page for the first time, they will load the cache info for all listed tokens in the background. The problem is, the whitelisted tokens do not load first, so it may take a loooong time for these to render even though there are only 2.

We want new users to have a fast UX. So, we load the cache info for these first.

Test Plan: `npm test`, `npm start` and open Cashtab in an incognito window, create a new wallet, navigate to Agora, notice the whitelisted orderbooks load in a few seconds (vs before a few minutes)

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D17697