Prevent occasional timeout in feature_assumevalid
Summary:
The header syncing can take some time on slow machines and offsets the timeout for the block syncing significantly, causing the test to fail to sync in time on some occasion.
This diff waits for the header to sync before starting the timeout which should make the test more robust.

Test Plan:
  ./test/functional/test_runner.py feature_assumevalid

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Differential Revision: https://reviews.bitcoinabc.org/D17713