[ecash-lib] Add `compressPk` to `Ecc`
Summary:
Sometimes we receive an uncompressed pubkey (e.g. the pubkey in the genesis tx, or in X509), but Ecc expects compressed pubkeys.

Therefore, we add `compressPk`. This is very simple to implement in pure TS (no WASM needed), but since the reverse operation (uncompressPk) would require WASM, we add it to Ecc for consistency.

Depends on D17732.

Test Plan: `npm test`

Reviewers: bytesofman, #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Subscribers: Fabien

Differential Revision: https://reviews.bitcoinabc.org/D17733