[avalanche] Fix a race in contenders voting test (again)
Summary: D17729 attempted to fix this race but erroneously assumes that once one peer gets a poll, all of them will have it. The better solution is to answer polls until we find the contenders we're looking for and then do the checks on that same poll request.

Test Plan:
Run the test a bunch of times:
  for I in {0..20}; do ./test/functional/test_runner.py abc_p2p_avalanche_contender_voting || break ; done

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D17737