[Cashtab] Validate agora accepts vs user balance
Summary:
Right now, Cashtab does not do this validation. So you can try to buy more than you can afford and expect an error at the tx building step.

This is lazy. We should validate for "can we afford this tx?" before we try to buy it.

We will improve this implementation going forward, but this is a key first step to supporting stuff like a "buy max" button.

The reason this was not implemented from the get go -- we need to calculate the actual price using agora methods after some validation is already done.

Test Plan: `npm test`

Reviewers: #bitcoin_abc, emack

Reviewed By: #bitcoin_abc, emack

Differential Revision: https://reviews.bitcoinabc.org/D17739