[ecash-coinselect, token-server] Remove ecash-coinselect from the monorepo
Summary:
Should have been done when it was deprecated from token-server. Looks like it is still in the token-server package-lock.json as an artifact. Clean up. Remove.

The logic from ecash-coinselect is obsolete and does not work with ecash-lib. Will be implemented in upcoming wallet lib.

Test Plan: `npm test`

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D17752