[electrum] Add native eCash support for compatible Trezor One firmware
Summary: The 1.13.0 firmware for the Trezor One introduced support for eCash derivation path, address and amounts. Let's use it if possible.

Test Plan:
With an up-to-date Trezor 1, open an existing wallet and check it still works in the BCH compatibility mode.
Create a new wallet, check it defaults to derivation 899. Send a tx and check the device screen properly shows XEC amounts and eCash addresses.

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Differential Revision: https://reviews.bitcoinabc.org/D17770