[token-server] Soften recaptcha check
Summary:
These days I get "are you a bot?" msg pretty much all the time. in the logs, legit requests are getting 0.5 scores. So, accept them.

It's pretty hard to believe we are getting so many "legit" claims these days when almost every time I try personally, I get recaptchad. But most of our claims are for CACHET, which there is not really a bot incentive to chase.

So, we're ok taking some attacks here, it is at any rate good data.

Test Plan: `npm test`

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Differential Revision: https://reviews.bitcoinabc.org/D17807