[electrum] properly exclude ALP tokens when sending XEC
Summary:
D17777 actually didn't achieve its goal of making it impossible to accidentaly burn ALP tokens, it only froze the coins in the Coins tab and made it impossible to spend them from there.

We need to filter the coins in the wallet's `get_utxos` method as well.

Also filter them in CashFusion (both when autofusing the entire wallet and when inputing coins from the address tab)

Depends on D17806

Test Plan:
In a wallet that has ALP tokens and SLP tokens, click the Max button in the spend tab, add a destination address, click "Preview" and check that none of the token UTXOs was included in the transaction.

Go to the Address tab, right click on an address that has only ALP tokens (no regular spendable utxos), check that there is no "Input coin to CashFusion" context action

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Subscribers: Fabien

Differential Revision: https://reviews.bitcoinabc.org/D17808