[avalanche] Cleanup circular dependency between peer manager and contender cache
Summary:
Now that peerManager owns the contender cache, it also makes sense for it to own the contender promotion logic as it pertains to proof status.

The unit test changes ensure we do not regress on the case where dangling remote proofs can still be accepted stake contenders.

Test Plan:
  ninja check check-functional

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D17813